Threads created

+++ srcs/utilitas.c
@@ -6,13 +6,13 @@
 /*   By: mrojas-e <mrojas-e@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/21 16:44:07 by mrojas-e          #+#    #+#             */
-/*   Updated: 2022/02/21 19:47:49 by mrojas-e         ###   ########.fr       */
+/*   Updated: 2022/02/22 13:34:04 by mrojas-e         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "sophos.h"
 
-unsigned long	get_time(ssize_t begin_time)
+unsigned long	get_time(int begin_time)
 {
 	struct timeval	t1;
